+++ modules/system_update.sh
@@ -10,7 +10,6 @@ fi
 
 if [[ "$AUTO_UPDATE" == "yes" ]]; then
   apt-get update \
-    && apt-get -y upgrade \
     && apt-get -y dist-upgrade \
     && apt-get -y autoremove \
     && apt-get -y autoclean